Add user removal from array and confirm alert

+++ UMS/src/app/users/users.component.ts
@@ -21,11 +21,17 @@ export class UsersComponent implements OnInit {
     }
 
     onDeleteUser(user: User) {
-        this.service.deleteUser(user).subscribe(
-            response => {
-                    alert(response['message']);
-            }
-        );
+       const deleteUser = confirm('Do you really want to delete user ' + user.name + '' + ' ' + user.lastname + '');
+       if (deleteUser) {
+         this.service.deleteUser(user).subscribe(
+           response => {
+
+             const idx = this.users.indexOf(user);
+             this.users.splice(idx, 1);
+             alert(response['message']);
+           }
+         );
+       }
     }
     onSelectUser(user: User) {
         const userCopy = Object.assign({}, user);
